adds calculator testing
